requirements.txt
colorama
docker
flasgger
flask
flask_caching
gunicorn
markdown-it-py
psutil
PyYAML
pyzipper
requests
stdlib_list
tabulate
